requirements.txt
paho-mqtt==1.5.1
requests==2.25.1
